requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-http
apache-airflow-providers-postgres
minio
python-dateutil
psycopg2-binary
apache-airflow-providers-microsoft-mssql
apache-airflow-providers-common-sql[pandas]
pyodbc
pyarrow
pandas
tqdm
numpy
pyjwt